Improvement

+++ app/controllers/product_controller.rb
@@ -3,8 +3,8 @@
 class ProductController < ApplicationController
   expose :products, parent: :current_user
   expose :product_for_sale, -> { Product.for_sale }
-  expose :product, parent: :current_user, build_params: :thing_params
-
+  expose :product, parent: :current_user
+  skip_before_action :authenticate_user!, only: [:index, :show]
   def create
     product.save
     render_item(product)
@@ -15,22 +15,25 @@ class ProductController < ApplicationController
   end
 
   def update
-    product.update(thing_params)
+    authorize(product)
+    product.update(product_params)
     render_item(product)
   end
 
   def destroy
+    authorize(product)
     product.destroy
     render_item(product)
   end
 
   def show
+    product = Product.find(params[:id])
     render_item(product)
   end
 
   private
 
-    def thing_params
+    def product_params
       params.permit(:name, :price, :description, :image, :category_id)
     end
 end
